+++ src/utils/talespire.js
@@ -32,8 +32,6 @@ const checkModifierKeys = (event, name) => {
 };
 
 export const triggerTalespire = (name, dice, extraDice) => {
-  dice = dice.replace(/d100/g, "d100+d10");
-
   let uri;
   if (typeof name === "string") {
     uri = `talespire://dice/${encodeURIComponent(name)}:${dice}${extraDice ? `/${dice}` : ""}`;
@@ -57,6 +55,8 @@ export const triggerTalespire = (name, dice, extraDice) => {
 
 export const talespireLink = (elem, label, dice, diceLabel) => {
   label = label?.trim();
+  dice = dice.replace(/d100/g, "d100+d10").replace(/[−–]/g, "-");
+  diceLabel = diceLabel.replace(/[−–]/g, "-");
 
   const link = document.createElement("button");
   link.classList.add("integrated-dice__container");