+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.20.0 (2025-02-15)
 
 ### Bug fixes